engine/balances.go
package engine

import (
	"errors"
	"sort"
	"strconv"
	"strings"
	"time"

	"github.com/accurateproject/accurate/dec"
	"github.com/accurateproject/accurate/utils"
	"go.uber.org/zap"
)

// Can hold different units as seconds or monetary
type Balance struct {
	UUID           string          `bson:"uuid"` //system wide unique
	ID             string          `bson:"id"`   // account wide unique
	Value          *dec.Dec        `bson:"value"`
	Directions     utils.StringMap `bson:"directions"`
	ExpirationDate time.Time       `bson:"expiration_date"`
	Weight         float64         `bson:"weight"`
	DestinationIDs utils.StringMap `bson:"destination_ids"`
	RatingSubject  string          `bson:"rating_subject"`
	Categories     utils.StringMap `bson:"categories"`
	SharedGroups   utils.StringMap `bson:"shared_groups"`
	Timings        []*RITiming     `bson:"timings"`
	TimingIDs      utils.StringMap `bson:"timing_ids"`
	Disabled       bool            `bson:"disabled"`
	Factor         ValueFactor     `bson:"factor"`
	Blocker        bool            `bson:"blocker"`
	Unlimited      bool            `bson:"unlimited"`
	precision      int
	account        *Account // used to store ub reference for shared balances
	dirty          bool
}

func (b *Balance) Equal(o *Balance) bool {
	if len(b.DestinationIDs) == 0 {
		b.DestinationIDs = utils.StringMap{utils.ANY: true}
	}
	if len(o.DestinationIDs) == 0 {
		o.DestinationIDs = utils.StringMap{utils.ANY: true}
	}
	return b.UUID == o.UUID &&
		b.ID == o.ID &&
		b.ExpirationDate.Equal(o.ExpirationDate) &&
		b.Weight == o.Weight &&
		b.DestinationIDs.Equal(o.DestinationIDs) &&
		b.Directions.Equal(o.Directions) &&
		b.RatingSubject == o.RatingSubject &&
		b.Categories.Equal(o.Categories) &&
		b.SharedGroups.Equal(o.SharedGroups) &&
		b.Disabled == o.Disabled &&
		b.Blocker == o.Blocker
}

// the default balance has standard Id
func (b *Balance) IsDefault() bool {
	return b.ID == utils.META_DEFAULT
}

func (b *Balance) IsExpired() bool {
	// check if it expires in the next second
	return !b.ExpirationDate.IsZero() && b.ExpirationDate.Before(time.Now().Add(1*time.Second))
}

func (b *Balance) IsActive() bool {
	return b.IsActiveAt(time.Now())
}

func (b *Balance) IsActiveAt(t time.Time) bool {
	if b.Disabled {
		return false
	}
	if len(b.Timings) == 0 {
		return true
	}
	for _, tim := range b.Timings {
		if tim.IsActiveAt(t) {
			return true
		}
	}
	return false
}

func (b *Balance) MatchCategory(category string) bool {
	return len(b.Categories) == 0 || b.Categories[category] == true
}

func (b *Balance) HasDestination() bool {
	return len(b.DestinationIDs) > 0 && b.DestinationIDs[utils.ANY] == false
}

func (b *Balance) HasDirection() bool {
	return len(b.Directions) > 0
}

func (b *Balance) MatchDestination(destinationID string) bool {
	return !b.HasDestination() || b.DestinationIDs[destinationID] == true
}

func (b *Balance) MatchActionTrigger(at *ActionTrigger) (bool, error) {
	filter, err := at.getFilter()
	if err != nil {
		return false, err
	}

	return filter.Query(b, false)
}

func (b *Balance) Clone() *Balance {
	if b == nil {
		return nil
	}
	n := &Balance{
		UUID:           b.UUID,
		ID:             b.ID,
		Value:          dec.New().Set(b.GetValue()), // this value is in seconds
		ExpirationDate: b.ExpirationDate,
		Weight:         b.Weight,
		RatingSubject:  b.RatingSubject,
		Categories:     b.Categories,
		SharedGroups:   b.SharedGroups,
		TimingIDs:      b.TimingIDs,
		Timings:        b.Timings, // should not be a problem with aliasing
		Blocker:        b.Blocker,
		Disabled:       b.Disabled,
		dirty:          b.dirty,
	}
	if b.DestinationIDs != nil {
		n.DestinationIDs = b.DestinationIDs.Clone()
	}
	if b.Directions != nil {
		n.Directions = b.Directions.Clone()
	}
	return n
}

func (b *Balance) getMatchingPrefixAndDestID(dest string) (prefix, destID string) {
	if len(b.DestinationIDs) != 0 && b.DestinationIDs[utils.ANY] == false {
		if dests, err := ratingStorage.GetDestinations(b.account.Tenant, dest, "", utils.DestMatching, utils.CACHED); err == nil {
			for _, dest := range dests {
				if b.DestinationIDs[dest.Name] == true {
					return dest.Code, dest.Name
				}
			}
		}
	}
	return
}

// Returns the available number of seconds for a specified credit
func (b *Balance) GetMinutesForCredit(origCD *CallDescriptor, initialCredit *dec.Dec) (duration time.Duration, credit *dec.Dec) {
	cd := origCD.Clone()
	availableDuration := time.Duration(b.GetValue().Int64()) * time.Second
	duration = availableDuration
	credit = dec.New()
	credit.Set(initialCredit)
	cc, err := b.GetCost(cd, false)
	if err != nil {
		utils.Logger.Error("Error getting new cost for balance subject: ", zap.String("tenant", cd.Tenant), zap.String("subject", b.RatingSubject), zap.Error(err))
		return 0, credit
	}
	if cc.deductConnectFee {
		connectFee := cc.GetConnectFee()
		if connectFee.Cmp(credit) <= 0 {
			credit.SubS(connectFee)
			// remove connect fee from the total cost
			cc.GetCost().SubS(connectFee)
		} else {
			return 0, credit
		}
	}
	if cc.GetCost().GtZero() {
		duration = 0
		for _, ts := range cc.Timespans {
			ts.createIncrementsSlice()
			if cd.MaxRate > 0 && cd.MaxRateUnit > 0 {
				rate, _, rateUnit := ts.RateInterval.GetRateParameters(ts.GetGroupStart())
				if rate.QuoS(dec.NewFloat(rateUnit.Seconds())).Cmp(dec.NewFloat(cd.MaxRate/cd.MaxRateUnit.Seconds())) > 0 {
					return
				}
			}
			ts.Increments.Reset()
			for _, incr := ts.Increments.Next(); incr != nil; _, incr = ts.Increments.Next() {
				if incr.getCost().Cmp(credit) <= 0 && availableDuration-incr.Duration >= 0 {
					credit.SubS(incr.Cost)
					duration += incr.Duration
					availableDuration -= incr.Duration
				} else {
					return
				}
			}
		}
	}
	return
}

// Gets the cost using balance RatingSubject if present otherwize
// retuns a callcost obtained using standard rating
func (b *Balance) GetCost(cd *CallDescriptor, getStandardIfEmpty bool) (*CallCost, error) {
	// testing only
	if cd.testCallcost != nil {
		return cd.testCallcost, nil
	}
	if b.RatingSubject != "" && !strings.HasPrefix(b.RatingSubject, utils.ZERO_RATING_SUBJECT_PREFIX) {
		origSubject := cd.Subject
		cd.Subject = b.RatingSubject
		origAccount := cd.Account
		cd.Account = cd.Subject
		cd.RatingInfos = nil
		cc, err := cd.getCost()
		// restor orig values
		cd.Subject = origSubject
		cd.Account = origAccount
		return cc, err
	}
	if getStandardIfEmpty {
		cd.RatingInfos = nil
		return cd.getCost()
	} else {
		cc := cd.CreateCallCost()
		cc.GetCost().Set(dec.Zero)
		return cc, nil
	}
}

func (b *Balance) GetValue() *dec.Dec {
	if b.Value == nil {
		b.Value = dec.New()
	}
	return b.Value
}

func (b *Balance) AddValue(amount *dec.Dec) {
	if b.Unlimited {
		return
	}
	b.SetValue(b.GetValue().AddS(amount))
}

func (b *Balance) SubstractValue(amount *dec.Dec) {
	if b.Unlimited {
		return
	}
	b.SetValue(b.GetValue().SubS(amount))
}

func (b *Balance) SetValue(amount *dec.Dec) {
	if b.Value == nil {
		b.Value = dec.New()
	}
	b.Value.Set(amount)
	//b.Value.Round(globalRoundingDecimals)
	b.dirty = true
}

func (b *Balance) SetDirty() {
	b.dirty = true
}

func (b *Balance) debitUnits(cd *CallDescriptor, ub *Account, moneyBalances Balances, count bool, dryRun, debitConnectFee bool) (cc *CallCost, err error) {
	if !b.IsActiveAt(cd.TimeStart) || b.GetValue().LteZero() {
		return
	}
	if duration, err := utils.ParseZeroRatingSubject(b.RatingSubject); err == nil {
		// we have *zero based units
		cc = cd.CreateCallCost()
		cc.Timespans = append(cc.Timespans, &TimeSpan{
			TimeStart: cd.TimeStart,
			TimeEnd:   cd.TimeEnd,
		})

		ts := cc.Timespans[0]
		ts.RoundToDuration(duration)
		ts.RateInterval = &RateInterval{
			Rating: &RIRate{
				Rates: RateGroups{
					&RateInfo{
						GroupIntervalStart: 0,
						Value:              dec.New(),
						RateIncrement:      duration,
						RateUnit:           duration,
					},
				},
			},
		}
		prefix, destid := b.getMatchingPrefixAndDestID(cd.Destination)
		if prefix == "" {
			prefix = cd.Destination
		}
		if destid == "" {
			destid = utils.ANY
		}
		ts.setRatingInfo(&RatingInfo{
			MatchedSubject: b.UUID,
			MatchedPrefix:  prefix,
			MatchedDestID:  destid,
			RatingPlanID:   utils.META_NONE,
		})
		ts.createIncrementsSlice()
		ts.Increments.Reset()
		for incIndex, inc := ts.Increments.Next(); inc != nil; incIndex, inc = ts.Increments.Next() {
			//log.Printf("INC: %s", utils.ToIJSON(inc))
			amount := dec.NewFloat(inc.Duration.Seconds())
			if b.Factor != nil {
				amount.QuoS(dec.NewFloat(b.Factor.GetValue(cd.TOR)))
			}
			//log.Print("B: ", utils.ToIJSON(b))
			if b.Unlimited || b.GetValue().Cmp(amount) >= 0 {
				b.SubstractValue(amount)
				if inc.BalanceInfo.Unit == nil {
					inc.BalanceInfo.Unit = &UnitInfo{
						UUID:          b.UUID,
						ID:            b.ID,
						DestinationID: cc.Destination,
						TOR:           cc.TOR,
						RateInterval:  nil,
					}
				}
				inc.BalanceInfo.Unit.getValue().Set(b.GetValue())
				inc.BalanceInfo.Unit.Consumed = amount.String()
				inc.BalanceInfo.AccountID = ub.Name
				inc.getCost().Set(dec.Zero)
				inc.paid++
				if count {
					pats := ub.countUnits(amount, cc.TOR, cc, b)
					inc.AddPostATIDs(incIndex, pats)
				}
			} else {
				// delete the rest of the unpiad increments/timespans
				if incIndex == 1 {
					// cat the entire current timespan
					cc.Timespans = nil
				} else {
					ts.SplitByIncrement(incIndex)
				}
				if len(cc.Timespans) == 0 {
					cc = nil
				}
				return cc, nil
			}
		}
	} else {
		// get the cost from balance
		//log.Printf("::::::: %+v", cd)
		cc, err = b.GetCost(cd, true)
		if err != nil {
			return nil, err
		}
		//log.Printf("CC: %s", utils.ToIJSON(cc))
		if debitConnectFee {
			// this is the first add, debit the connect fee
			if ub.DebitConnectionFee(cc, moneyBalances, count, true) == false {
				// found blocker balance
				return nil, nil
			}
		}
		cc.Timespans.Decompress()
		//log.Printf("CC: %s", utils.ToIJSON(cc))

		for tsIndex, ts := range cc.Timespans {
			if ts.Increments == nil {
				ts.createIncrementsSlice()
			}

			if ts.RateInterval == nil {
				utils.Logger.Error("Nil RateInterval ERROR on ", zap.Any("TS", ts), zap.Any("CC", cc), zap.Any("CD", cd))
				return nil, errors.New("timespan with no rate interval assigned")
			}
			maxCost, strategy := ts.RateInterval.GetMaxCost()
			ts.Increments.Reset()
			for incIndex, inc := ts.Increments.Next(); inc != nil; incIndex, inc = ts.Increments.Next() {
				// debit minutes and money
				amount := dec.NewFloat(inc.Duration.Seconds())
				if b.Factor != nil {
					amount.QuoS(dec.NewFloat(b.Factor.GetValue(cd.TOR)))
				}
				cost := inc.Cost
				inc.paid++
				if strategy == utils.MAX_COST_DISCONNECT && cd.GetMaxCostSoFar().Cmp(maxCost) >= 0 {
					// cat the entire current timespan
					cc.maxCostDisconect = true
					if dryRun {
						if incIndex == 1 {
							// cat the entire current timespan
							cc.Timespans = cc.Timespans[:tsIndex]
						} else {
							ts.SplitByIncrement(incIndex)
							cc.Timespans = cc.Timespans[:tsIndex+1]
						}
						return cc, nil
					}
				}
				if strategy == utils.MAX_COST_FREE && cd.GetMaxCostSoFar().Cmp(maxCost) >= 0 {
					cost = dec.New()
					ts.Increments.MaxCostFreeIndex = incIndex - 1
					if inc.BalanceInfo.Monetary == nil {
						inc.BalanceInfo.Monetary = &MonetaryInfo{
							UUID:         b.UUID,
							ID:           b.ID,
							RateInterval: ts.RateInterval,
						}
					}
					inc.BalanceInfo.Monetary.getValue().Set(b.GetValue())
					inc.BalanceInfo.AccountID = ub.Name
					inc.paid = inc.CompressFactor
					if count {
						pats := ub.countUnits(cost, utils.MONETARY, cc, b)
						inc.AddPostATIDs(incIndex, pats)
					}
					// go to nextincrement
					break
				}
				var moneyBal *Balance
				for _, mb := range moneyBalances {
					if mb.Unlimited || mb.GetValue().Cmp(cost) >= 0 {
						moneyBal = mb
						break
					}
				}
				if (cost.IsZero() || moneyBal != nil) && (b.Unlimited || b.GetValue().Cmp(amount) >= 0) {
					//log.Print("INC: ", utils.ToIJSON(inc), amount)
					b.SubstractValue(amount)
					if inc.BalanceInfo.Unit == nil {
						inc.BalanceInfo.Unit = &UnitInfo{
							UUID:          b.UUID,
							ID:            b.ID,
							DestinationID: cc.Destination,
							TOR:           cc.TOR,
							RateInterval:  ts.RateInterval,
						}
					}
					inc.BalanceInfo.Unit.getValue().Set(b.GetValue())
					inc.BalanceInfo.Unit.Consumed = amount.String()
					inc.BalanceInfo.AccountID = ub.Name
					if !cost.IsZero() {
						moneyBal.SubstractValue(cost)
						if inc.BalanceInfo.Monetary == nil {
							inc.BalanceInfo.Monetary = &MonetaryInfo{
								UUID: moneyBal.UUID,
								ID:   moneyBal.ID,
							}
						}
						inc.BalanceInfo.Monetary.getValue().Set(moneyBal.GetValue())
						cd.GetMaxCostSoFar().AddS(cost)
					}
					inc.paid++
					if count {
						pats := ub.countUnits(amount, cc.TOR, cc, b)
						inc.AddPostATIDs(incIndex, pats)
						if !cost.IsZero() {
							pats = ub.countUnits(cost, utils.MONETARY, cc, moneyBal)
							inc.AddPostATIDs(incIndex, pats)
						}
					}
				} else {
					// delete the rest of the unpiad increments/timespans
					if incIndex == 1 {
						// cat the entire current timespan
						cc.Timespans = cc.Timespans[:tsIndex]
					} else {
						ts.SplitByIncrement(incIndex)
						cc.Timespans = cc.Timespans[:tsIndex+1]
					}
					if len(cc.Timespans) == 0 {
						cc = nil
					}
					return cc, nil
				}
			}
		}
	}
	return
}

func (b *Balance) debitMoney(cd *CallDescriptor, ub *Account, moneyBalances Balances, count bool, dryRun, debitConnectFee bool) (cc *CallCost, err error) {
	if !b.IsActiveAt(cd.TimeStart) || b.GetValue().LteZero() {
		return
	}
	//log.Print("B: ", utils.ToIJSON(b))
	//log.Printf("}}}}}}} %+v", cd.testCallcost)
	//log.Print("CD: ", utils.ToIJSON(cd))
	cc, err = b.GetCost(cd, true)
	if err != nil {
		return nil, err
	}
	//log.Print("cc: " + utils.ToIJSON(cc))
	if debitConnectFee {
		// this is the first add, debit the connect fee
		if ub.DebitConnectionFee(cc, moneyBalances, count, true) == false {
			// balance is blocker
			return nil, nil
		}
	}
	cc.Timespans.Decompress()
	//log.Printf("CallCost In Debit: %+v", utils.ToIJSON(cc))
	//for _, ts := range cc.Timespans {
	//	log.Printf("CC_TS: %+v", ts.RateInterval.Rating.Rates[0])
	//}
	for tsIndex, ts := range cc.Timespans {
		if ts.Increments == nil {
			ts.createIncrementsSlice()
		}
		//log.Printf("TS: %s", utils.ToIJSON(ts))
		if ts.RateInterval == nil {
			utils.Logger.Error("Nil RateInterval ERROR on ", zap.Any("TS", ts), zap.Any("CC", cc), zap.Any("CD", cd))
			return nil, errors.New("timespan with no rate interval assigned")
		}
		maxCost, strategy := ts.RateInterval.GetMaxCost()
		//log.Printf("Timing: %+v", ts.RateInterval.Timing)
		//log.Printf("Rate: %+v", ts.RateInterval.Rating)
		ts.Increments.Reset()
		for incIndex, inc := ts.Increments.Next(); inc != nil; incIndex, inc = ts.Increments.Next() {
			// check standard subject tags
			//log.Print("INC: ", utils.ToIJSON(inc))
			amount := inc.Cost
			if strategy == utils.MAX_COST_DISCONNECT && cd.GetMaxCostSoFar().Cmp(maxCost) >= 0 {
				// cat the entire current timespan
				cc.maxCostDisconect = true
				if dryRun {
					if incIndex == 1 {
						// cat the entire current timespan
						cc.Timespans = cc.Timespans[:tsIndex]
					} else {
						ts.SplitByIncrement(incIndex)
						cc.Timespans = cc.Timespans[:tsIndex+1]
					}
					return cc, nil
				}
			}
			if strategy == utils.MAX_COST_FREE && cd.GetMaxCostSoFar().Cmp(maxCost) >= 0 {
				amount = dec.New()
				ts.Increments.MaxCostFreeIndex = incIndex - 1
				if inc.BalanceInfo.Monetary == nil {
					inc.BalanceInfo.Monetary = &MonetaryInfo{
						UUID: b.UUID,
						ID:   b.ID,
					}
				}
				inc.BalanceInfo.Monetary.getValue().Set(b.GetValue())
				inc.BalanceInfo.AccountID = ub.Name
				if b.RatingSubject != "" {
					inc.BalanceInfo.Monetary.RateInterval = ts.RateInterval
				}
				inc.paid = inc.CompressFactor
				if count {
					pats := ub.countUnits(amount, utils.MONETARY, cc, b)
					inc.AddPostATIDs(incIndex, pats)
				}
				//log.Printf("TS: %+v", cc.Cost)
				// go to nextincrement
				break
			}
			if b.Unlimited || b.GetValue().Cmp(amount) >= 0 {
				b.SubstractValue(amount)
				cd.GetMaxCostSoFar().AddS(amount)
				if inc.BalanceInfo.Monetary == nil {
					inc.BalanceInfo.Monetary = &MonetaryInfo{
						UUID: b.UUID,
						ID:   b.ID,
					}
				}
				inc.BalanceInfo.Monetary.getValue().Set(b.GetValue())
				inc.BalanceInfo.AccountID = ub.Name
				if b.RatingSubject != "" {
					inc.BalanceInfo.Monetary.RateInterval = ts.RateInterval
				}
				inc.paid++
				if count {
					pats := ub.countUnits(amount, utils.MONETARY, cc, b)
					inc.AddPostATIDs(incIndex, pats)
				}
			} else {
				// delete the rest of the unpiad increments/timespans
				if incIndex == 1 {
					// cat the entire current timespan
					cc.Timespans = cc.Timespans[:tsIndex]
				} else {
					ts.SplitByIncrement(incIndex)
					cc.Timespans = cc.Timespans[:tsIndex+1]
				}
				if len(cc.Timespans) == 0 {
					cc = nil
				}
				return cc, nil
			}
		}
	}
	//log.Printf("END: %+v", cd.testCallcost)
	if len(cc.Timespans) == 0 {
		cc = nil
	}
	return cc, nil
}

// Converts the balance towards compressed information to be displayed
func (b *Balance) AsBalanceSummary(typ string) *BalanceSummary {
	bd := &BalanceSummary{ID: b.ID, Type: typ, Value: b.GetValue().String(), Disabled: b.Disabled}
	if bd.ID == "" {
		bd.ID = b.UUID
	}
	return bd
}

/*
Structure to store minute buckets according to weight, precision or price.
*/
type Balances []*Balance

func (bc Balances) Sort() {
	// we need the better ones at the beginning
	sort.Slice(bc, func(j, i int) bool {
		return bc[i].precision < bc[j].precision ||
			(bc[i].precision == bc[j].precision && bc[i].Weight < bc[j].Weight)
	})
}

func (bc Balances) GetTotalValue() (total *dec.Dec) {
	total = dec.New()
	for _, b := range bc {
		if !b.IsExpired() && b.IsActive() {
			total.AddS(b.GetValue())
		}
	}
	//total.Round(globalRoundingDecimals)
	return
}

func (bc Balances) Equal(o Balances) bool {
	if len(bc) != len(o) {
		return false
	}
	bc.Sort()
	o.Sort()
	for i := 0; i < len(bc); i++ {
		if !bc[i].Equal(o[i]) {
			return false
		}
	}
	return true
}

func (bc Balances) Clone() Balances {
	var newChain Balances
	for _, b := range bc {
		newChain = append(newChain, b.Clone())
	}
	return newChain
}

func (bc Balances) GetBalance(uuid string) *Balance {
	for _, balance := range bc {
		if balance.UUID == uuid {
			return balance
		}
	}
	return nil
}

func (bc Balances) HasBalance(balance *Balance) bool {
	for _, b := range bc {
		if b.Equal(balance) {
			return true
		}
	}
	return false
}

func (bc Balances) SaveDirtyBalances(acc *Account) {
	savedAccounts := make(map[string]bool)
	for _, b := range bc {
		if b.dirty {
			b.GetValue().Round(globalRoundingDecimals)
			// publish event
			accID := ""
			allowNegative := ""
			disabled := ""
			if b.account != nil { // only publish modifications for balances with account set
				//utils.LogStack()
				accID = b.account.Name
				allowNegative = strconv.FormatBool(b.account.AllowNegative)
				disabled = strconv.FormatBool(b.account.Disabled)
				Publish(CgrEvent{
					"EventName":            utils.EVT_ACCOUNT_BALANCE_MODIFIED,
					"Uuid":                 b.UUID,
					"Id":                   b.ID,
					"Value":                b.GetValue().String(),
					"ExpirationDate":       b.ExpirationDate.String(),
					"Weight":               strconv.FormatFloat(b.Weight, 'f', -1, 64),
					"DestinationIDs":       b.DestinationIDs.String(),
					"Directions":           b.Directions.String(),
					"RatingSubject":        b.RatingSubject,
					"Categories":           b.Categories.String(),
					"SharedGroups":         b.SharedGroups.String(),
					"TimingIDs":            b.TimingIDs.String(),
					"Account":              accID,
					"AccountAllowNegative": allowNegative,
					"AccountDisabled":      disabled,
				})
			}
		}
		if b.account != nil && b.account != acc && b.dirty && savedAccounts[b.account.Name] == false {
			if err := accountingStorage.SetAccount(b.account); err != nil {
				utils.Logger.Error("< SaveDirtyBalances> err: ", zap.Error(err))
			}
			savedAccounts[b.account.Name] = true
		}
	}
}

type ValueFactor map[string]float64

func (f ValueFactor) GetValue(tor string) float64 {
	if value, ok := f[tor]; ok {
		return value
	}
	return 1.0
}

// BalanceSummary balance information
type BalanceSummary struct {
	ID       string // ID or UUID if not defined
	Type     string // *voice, *data, etc
	Value    string
	Disabled bool
}
